Add the Positional Encoding module (#3)
The purpose of positional encoding is to add information about the relative positions of tokens within a sequence since the self-attention mechanism does not have any inherent notion of order. This is important as a transformer doesn't otherwise inherently have any way of easily way of obtaining the order of tokens.